fin de la structure du backoffice et de la page dashboard

+++ backoffice/js/back-office.js
@@ -1,7 +1,12 @@
-var tableDesTitre = {
-    dashBoardSite : 'Dashboard',
-    blog : 'Blog'
-}
+var menuActif = '',
+    tableDesTitre = {
+        dashBoardSite : 'Dashboard',
+        blog : 'Blog',
+        evenement : 'Événement',
+        statistiques : 'Statistiques',
+        utilisateurs : 'Utilisateurs',
+        shop : 'Shop'
+    };
 
 window.onload = function(){
     ZoneBoite = new DomWork(byId('insertZone'));
@@ -12,10 +17,7 @@ window.onload = function(){
         changeOnglet('dashBoardSite');
         
         //initialise l'ouverture des menus
-        $(document).ready(function(){
-            console.log('ok');
-            $('.collapsible').collapsible();
-        });
+        $(document).ready(function(){ $('.collapsible').collapsible();});
         
     },1000);
     
@@ -34,43 +36,55 @@ function changeOnglet(cible){
         insertZone = byId('insertZone'),
         classBoite = byQueryAll('.boite');
     
-    //supprime les anciennes boites
-    for(var i = 0 ; i < classBoite.length ; i++) classBoite[i].deleteNode();
-    
-    //affiche le titre
-    titreZone.textContent = tableDesTitre[cible];
-    
-    insertZone.innerHTML = '';
-    
-    //selectionne les vus à montrer
-    if(cible == 'dashBoardSite'){
-        ZoneBoite.insertDomNode([
-            {type : 'div' , attributes : {class : 'row stretchCol'}, contents : [
-                {type : 'div', attributes : {class : 'col s9'}, contents : structureBoite('equalizer', 'Vue de l\'activité', 'boiteAjax/vuActiviteSite.html')},
-                {type : 'div', attributes : {class : 'col s3'}, contents : structureBoite('fast_forward', 'Actions rapides', 'boiteAjax/ActionRapideSite.html')}
-            ]},
-            {type : 'div' , attributes : {class : 'row stretchCol'}, contents : [
-                {type : 'div', attributes : {class : 'col s5'}, contents : structureBoite('mode_edit', 'Activitées éditorial', 'boiteAjax/ActiviteEditorial.html')},
-                {type : 'div', attributes : {class : 'col s7'}, contents : structureBoite('star', 'Derniers avis sur les produits', 'boiteAjax/DernierAvisProduits.html')}
-            ]}
-        ]);
+    if(menuActif != cible){
+        
+        menuActif = cible;
+        
+        //supprime les anciennes boites
+        for(var i = 0 ; i < classBoite.length ; i++) classBoite[i].deleteNode();
+
+        //affiche le titre
+        titreZone.textContent = tableDesTitre[cible];
+
+        insertZone.innerHTML = '';
+
+        //selectionne les vus à montrer
+        if(cible == 'dashBoardSite'){
+            ZoneBoite.insertDomNode([
+                {type : 'div' , attributes : {class : 'row stretchCol'}, contents : [
+                    {type : 'div', attributes : {class : 'col s9'}, contents : structureBoite('equalizer', 'Vue de l\'activité', 'boiteAjax/vuActiviteSite.html', creatDashSiteData)},
+                    {type : 'div', attributes : {class : 'col s3'}, contents : structureBoite('fast_forward', 'Actions rapides', 'boiteAjax/ActionRapideSite.html')}
+                ]},
+                {type : 'div' , attributes : {class : 'row stretchCol'}, contents : [
+                    {type : 'div', attributes : {class : 'col s5'}, contents : structureBoite('mode_edit', 'Activitées éditorial', 'boiteAjax/ActiviteEditorial.html')},
+                    {type : 'div', attributes : {class : 'col s7'}, contents : structureBoite('star', 'Derniers avis sur les produits', 'boiteAjax/DernierAvisProduits.html')}
+                ]}
+            ]);
+        }
+        if(cible == 'blog'){
+             ZoneBoite.insertDomNode([
+                {type : 'div', attributes : {class : 'col s12'}, contents : structureBoite('message', 'Les blogs', 'boiteAjax/lesBlogs.html')}
+            ]);
+        }
     }
-    if(cible == 'blog'){
-         ZoneBoite.insertDomNode([
-            {type : 'div', attributes : {class : 'col s12'}, contents : structureBoite('message', 'Les blogs', 'boiteAjax/lesBlogs.html')}
-        ]);
-    }
-    
 }
 
-function structureBoite(icone, titre, path){
+function structureBoite(icone, titre, path, onload = null){
     return [
         {type : 'div', attributes : {class : 'boiteAjax card'}, contents : [
             {type : 'div', attributes : {class : 'enteteBoite'}, contents : [
                 {type : 'i', attributes : {class : 'small material-icons colorIcone'}, contents : icone},
                 {type : 'span', attributes : {class : 'titreDeLaBoite'}, contents : titre}
             ]},
-            {type : 'div', attributes : {class : 'divInnerBoite'}, contents : [{ajax : {path : path}}]}
+            {type : 'div', attributes : {class : 'divInnerBoite'}, contents : [{ajax : {path : path, onload : onload}}]}
         ]}
     ]
+}
+
+function selectedMenu(cible){
+    var btMenu = byClass('collapsible-header selectedMenu'),
+        selectedItem = byId(cible);
+    
+    btMenu[0].className = btMenu[0].className.replace('selectedMenu','');
+    selectedItem.className += ' selectedMenu';    
 }
